+++ tests/testthat/test_filter_ref_ranges.R
@@ -1,9 +1,5 @@
 test_that("Sample filtering tests", {
-    hVcfFileDir <- system.file("extdata", package = "rPHG2")
-    hVcfFiles   <- list.files(hVcfFileDir, pattern = ".h.vcf$", full.names = TRUE)
-    locCon      <- PHGLocalCon(hVcfFiles)
-    graph       <- buildHaplotypeGraph(locCon)
-    pds         <- readPhgDataSet(graph)
+    pds <- makeExamplePds()
 
     gr01 <- GenomicRanges::GRanges(
         seqnames = c("1"),