+++ tests/testthat/test_read_hap_id_metadata.R
@@ -17,11 +17,7 @@ test_that("Hap ID metadata reading tests from JVM", {
 
 
 test_that("Hap ID metadata reading tests from PHGDataSet", {
-    hVcfFileDir <- system.file("extdata", package = "rPHG2")
-    hVcfFiles   <- list.files(hVcfFileDir, pattern = ".h.vcf$", full.names = TRUE)
-    locCon      <- PHGLocalCon(hVcfFiles)
-    graph       <- buildHaplotypeGraph(locCon)
-    pds         <- readPhgDataSet(graph)
+    pds <- makeExamplePds()
 
     obsHapIdMetaData <- readHapIdMetaData(pds)
     expect_true(is(obsHapIdMetaData, "data.frame"))