+++ tests/testthat/test_read_hap_id_pos_metadata.R
@@ -17,11 +17,7 @@ test_that("Hap ID positional metadata reading tests from JVM", {
 
 
 test_that("Hap ID positional metadata reading tests from PHGDataSet", {
-    hVcfFileDir <- system.file("extdata", package = "rPHG2")
-    hVcfFiles   <- list.files(hVcfFileDir, pattern = ".h.vcf$", full.names = TRUE)
-    locCon      <- PHGLocalCon(hVcfFiles)
-    graph       <- buildHaplotypeGraph(locCon)
-    pds         <- readPhgDataSet(graph)
+    pds <- makeExamplePds()
 
     obsHapIdPosMetaData <- readHapIdPosMetaData(pds)