+++ tests/testthat/test_read_hap_ids.R
@@ -13,11 +13,7 @@ test_that("Hap ID matrix reading tests from JVM", {
 
 
 test_that("Hap ID matrix reading tests from PHGDataSet", {
-    hVcfFileDir <- system.file("extdata", package = "rPHG2")
-    hVcfFiles   <- list.files(hVcfFileDir, pattern = ".h.vcf$", full.names = TRUE)
-    locCon      <- PHGLocalCon(hVcfFiles)
-    graph       <- buildHaplotypeGraph(locCon)
-    pds         <- readPhgDataSet(graph)
+    pds <- makeExamplePds()
 
     obsHapIds <- readHapIds(pds)