+++ NEWS.md
@@ -1,3 +1,13 @@
+## rPHG2 0.9
+* Added new generic function, `findSamplesByHaplotype()`
+  + Using `PHGDataSet` objects, users can return all given samples from a
+    provided collection of haplotype IDs.
+* Update `seqnames()` generic to `seqNames()`:
+  + This is to avoid naming collisions in `GenomeInfoDb`
+* Fixes to possible errors related to `ggplot2` v4.0
+  + More info can be found [here](https://tidyverse.org/blog/2025/09/ggplot2-4-0-0/#adopting-s7).
+
+
 ## rPHG2 0.8
 * Fixed missing IDs returned in `readHapIds()` method calls:
   + missing values (i.e., `"null"`) now return as `NA`s in the matrix object